requirements.txt
# Core AI providers
google-generativeai>=0.3.0
groq>=0.4.0
openai>=1.10.0

# Trend analysis
pytrends>=4.9.0

# HTTP & utilities
requests>=2.31.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
pydantic>=2.5.0

# Media services
pexels-api>=1.0.1
youtube-search-python>=1.6.6

# Content processing
markdown>=3.5.0
beautifulsoup4>=4.12.0
Pillow>=10.0.0

# CLI and logging
rich>=13.7.0
